refactoring doe
