+++ bofire/strategies/doe/design.py
@@ -1,7 +1,4 @@
-import time
 import warnings
-from itertools import combinations_with_replacement, product
-from queue import PriorityQueue
 from typing import Dict, List, Optional, Tuple, Union
 
 import numpy as np
@@ -16,7 +13,6 @@ from bofire.data_models.constraints.api import (
 )
 from bofire.data_models.domain.api import Domain
 from bofire.data_models.enum import SamplingMethodEnum
-from bofire.data_models.features.api import ContinuousInput, Input
 from bofire.data_models.strategies.api import RandomStrategy as RandomStrategyDataModel
 from bofire.data_models.types import Bounds
 from bofire.strategies.doe.objective import get_objective_class
@@ -28,343 +24,6 @@ from bofire.strategies.doe.utils import (
 )
 from bofire.strategies.enum import OptimalityCriterionEnum
 from bofire.strategies.random import RandomStrategy
-
-
-def find_local_max_ipopt_BaB(
-    domain: Domain,
-    model_type: Union[str, Formula],
-    n_experiments: Optional[int] = None,
-    delta: float = 1e-7,
-    ipopt_options: Optional[Dict] = None,
-    sampling: Optional[pd.DataFrame] = None,
-    fixed_experiments: Optional[pd.DataFrame] = None,
-    partially_fixed_experiments: Optional[pd.DataFrame] = None,
-    objective: OptimalityCriterionEnum = OptimalityCriterionEnum.D_OPTIMALITY,
-    categorical_groups: Optional[List[List[ContinuousInput]]] = None,
-    discrete_variables: Optional[Dict[str, Tuple[ContinuousInput, List[float]]]] = None,
-    verbose: bool = False,
-    transform_range: Optional[Bounds] = None,
-) -> pd.DataFrame:
-    """Function computing a d-optimal design" for a given domain and model.
-    It allows for the problem to have categorical values which is solved by Branch-and-Bound
-        Args:
-            domain (Domain): domain containing the inputs and constraints.
-            model_type (str, Formula): keyword or formulaic Formula describing the model. Known keywords
-                are "linear", "linear-and-interactions", "linear-and-quadratic", "fully-quadratic".
-            n_experiments (int): Number of experiments. By default the value corresponds to
-                the number of model terms - dimension of ker() + 3.
-            delta (float): Regularization parameter. Default value is 1e-3.
-            ipopt_options (Dict, optional): options for IPOPT. For more information see [this link](https://coin-or.github.io/Ipopt/OPTIONS.html)
-            sampling (pd.DataFrame): dataframe containing the initial guess.
-            fixed_experiments (pd.DataFrame): dataframe containing experiments that will be definitely part of the design.
-                Values are set before the optimization.
-            partially_fixed_experiments (pd.DataFrame): dataframe containing (some) fixed variables for experiments.
-                Values are set before the optimization. Within one experiment not all variables need to be fixed.
-                Variables can be fixed to one value or can be set to a range by setting a tuple with lower and upper bound
-                Non-fixed variables have to be set to None or nan.
-            objective (OptimalityCriterionEnum): OptimalityCriterionEnum object indicating which objective function to use.
-            categorical_groups (Optional[List[List[ContinuousInput]]]). Represents the different groups of the
-               relaxed categorical variables. Defaults to None.
-            discrete_variables (Optional[Dict[str, Tuple[ContinuousInput, List[float]]]]): dict of relaxed discrete inputs
-                with key:(relaxed variable, valid values). Defaults to None
-            verbose (bool): if true, print information during the optimization process
-            transform_range (Optional[Bounds]): range to which the input variables are transformed.
-                If None is provided, the features will not be scaled. Defaults to None.
-        Returns:
-            A pd.DataFrame object containing the best found input for the experiments. In general, this is only a
-            local optimum.
-    """
-    from bofire.strategies.doe.branch_and_bound import NodeExperiment, bnb
-
-    if categorical_groups is None:
-        categorical_groups = []
-
-    model_formula = get_formula_from_string(
-        model_type=model_type, rhs_only=True, domain=domain
-    )
-
-    n_experiments = get_n_experiments(model_formula, n_experiments)
-
-    # get objective function
-    objective_class = get_objective_class(objective)
-    objective_class = objective_class(
-        domain=domain,
-        model=model_formula,
-        n_experiments=n_experiments,
-        delta=delta,
-        transform_range=transform_range,
-    )
-
-    # setting up initial node in the branch-and-bound tree
-    column_keys = domain.inputs.get_keys()
-
-    if fixed_experiments is not None:
-        subtract = len(fixed_experiments)
-        initial_branch = pd.DataFrame(
-            np.full((n_experiments - subtract, len(column_keys)), None),
-            columns=column_keys,
-        )
-        initial_branch = pd.concat([fixed_experiments, initial_branch]).reset_index(
-            drop=True
-        )
-    else:
-        initial_branch = pd.DataFrame(
-            np.full((n_experiments, len(column_keys)), None),
-            columns=column_keys,
-        )
-
-    if partially_fixed_experiments is not None:
-        partially_fixed_experiments = pd.concat(
-            [
-                partially_fixed_experiments,
-                pd.DataFrame(
-                    np.full(
-                        (
-                            n_experiments - len(partially_fixed_experiments),
-                            len(domain.inputs),
-                        ),
-                        None,
-                    ),
-                    columns=domain.inputs.get_keys(includes=Input),
-                ),
-            ]
-        ).reset_index(drop=True)
-
-        initial_branch.mask(
-            partially_fixed_experiments.notnull(),  # type: ignore
-            other=partially_fixed_experiments,
-            inplace=True,
-        )
-
-    initial_design = find_local_max_ipopt(
-        domain,
-        model_formula,
-        n_experiments,
-        delta,
-        ipopt_options,
-        sampling,
-        None,
-        partially_fixed_experiments=initial_branch,
-        objective=objective,
-    )
-    initial_value = objective_class.evaluate(
-        initial_design.to_numpy().flatten(),
-    )
-
-    initial_node = NodeExperiment(
-        initial_branch,
-        initial_design,
-        initial_value,
-        categorical_groups,
-        discrete_variables,
-    )
-
-    # initializing branch-and-bound queue
-    initial_queue = PriorityQueue()
-    initial_queue.put(initial_node)
-
-    # starting branch-and-bound
-    result_node = bnb(
-        initial_queue,
-        domain=domain,
-        model_type=model_formula,
-        n_experiments=n_experiments,
-        delta=delta,
-        ipopt_options=ipopt_options,
-        sampling=sampling,
-        fixed_experiments=None,
-        objective=objective,
-        verbose=verbose,
-    )
-
-    return result_node.design_matrix
-
-
-def find_local_max_ipopt_exhaustive(
-    domain: Domain,
-    model_type: Union[str, Formula],
-    n_experiments: Optional[int] = None,
-    delta: float = 1e-7,
-    ipopt_options: Optional[Dict] = None,
-    sampling: Optional[pd.DataFrame] = None,
-    fixed_experiments: Optional[pd.DataFrame] = None,
-    objective: OptimalityCriterionEnum = OptimalityCriterionEnum.D_OPTIMALITY,
-    partially_fixed_experiments: Optional[pd.DataFrame] = None,
-    categorical_groups: Optional[List[List[ContinuousInput]]] = None,
-    discrete_variables: Optional[Dict[str, Tuple[ContinuousInput, List[float]]]] = None,
-    verbose: bool = False,
-    transform_range: Optional[Bounds] = None,
-) -> pd.DataFrame:
-    """Function computing a d-optimal design" for a given domain and model.
-    It allows for the problem to have categorical values which is solved by exhaustive search
-        Args:
-            domain (Domain): domain containing the inputs and constraints.
-            model_type (str, Formula): keyword or formulaic Formula describing the model. Known keywords
-                are "linear", "linear-and-interactions", "linear-and-quadratic", "fully-quadratic".
-            n_experiments (int): Number of experiments. By default the value corresponds to
-                the number of model terms - dimension of ker() + 3.
-            delta (float): Regularization parameter. Default value is 1e-3.
-            ipopt_options (Dict, optional): options for IPOPT. For more information see [this link](https://coin-or.github.io/Ipopt/OPTIONS.html)
-            sampling (pd.DataFrame): dataframe containing the initial guess.
-            fixed_experiments (pd.DataFrame): dataframe containing experiments that will be definitely part of the design.
-                Values are set before the optimization.
-            objective (OptimalityCriterionEnum): OptimalityCriterionEnum object indicating which objective function to use.
-            partially_fixed_experiments (pd.DataFrame): dataframe containing (some) fixed variables for experiments.
-                Values are set before the optimization. Within one experiment not all variables need to be fixed.
-                Variables can be fixed to one value or can be set to a range by setting a tuple with lower and upper bound
-                Non-fixed variables have to be set to None or nan.
-            categorical_groups (Optional[List[List[ContinuousInput]]]). Represents the different groups of the
-               relaxed categorical variables. Defaults to None.
-            discrete_variables (Optional[Dict[str, Tuple[ContinuousInput, List[float]]]]): dict of relaxed discrete inputs
-                with key:(relaxed variable, valid values). Defaults to None
-            verbose (bool): if true, print information during the optimization process
-            transform_range (Optional[Bounds]): range to which the input variables are transformed.
-        Returns:
-            A pd.DataFrame object containing the best found input for the experiments. In general, this is only a
-            local optimum.
-    """
-
-    if categorical_groups is None:
-        categorical_groups = []
-
-    if discrete_variables is not None or len(discrete_variables) > 0:  # type: ignore
-        raise NotImplementedError(
-            "Exhaustive search for discrete variables is not implemented yet."
-        )
-
-    # get objective function
-    model_formula = get_formula_from_string(
-        model_type=model_type, rhs_only=True, domain=domain
-    )
-    objective_class = get_objective_class(objective)
-    objective_class = objective_class(
-        domain=domain,
-        model=model_formula,
-        n_experiments=n_experiments,
-        delta=delta,
-        transform_range=transform_range,
-    )
-
-    # get binary variables
-    binary_vars = [var for group in categorical_groups for var in group]
-    list_keys = [var.key for var in binary_vars]
-
-    # determine possible fixations of the different categories
-    allowed_fixations = []
-    for group in categorical_groups:
-        allowed_fixations.append(np.eye(len(group)))
-
-    n_experiments = get_n_experiments(model_formula, n_experiments)
-    n_non_fixed_experiments = n_experiments
-    if fixed_experiments is not None:
-        n_non_fixed_experiments -= len(fixed_experiments)
-
-    allowed_fixations = product(*allowed_fixations)
-    all_n_fixed_experiments = combinations_with_replacement(
-        allowed_fixations, n_non_fixed_experiments
-    )
-
-    if partially_fixed_experiments is not None:
-        partially_fixed_experiments = pd.concat(
-            [
-                partially_fixed_experiments,
-                pd.DataFrame(
-                    np.full(
-                        (
-                            n_non_fixed_experiments - len(partially_fixed_experiments),
-                            len(domain.inputs),
-                        ),
-                        None,
-                    ),
-                    columns=domain.inputs.get_keys(includes=Input),
-                ),
-            ]
-        ).reset_index(drop=True)
-
-    # testing all different fixations
-    column_keys = domain.inputs.get_keys()
-    group_keys = [var.key for group in categorical_groups for var in group]
-    minimum = float("inf")
-    optimal_design = pd.DataFrame()
-    all_n_fixed_experiments = list(all_n_fixed_experiments)
-    for i, binary_fixed_experiments in enumerate(all_n_fixed_experiments):
-        if verbose:
-            start_time = time.time()
-        # setting up the pd.Dataframe for the partially fixed experiment
-        binary_fixed_experiments = np.array(
-            [
-                var
-                for experiment in binary_fixed_experiments
-                for group in experiment
-                for var in group
-            ]
-        ).reshape(n_non_fixed_experiments, len(binary_vars))
-
-        binary_fixed_experiments = pd.DataFrame(
-            binary_fixed_experiments, columns=group_keys
-        )
-        one_set_of_experiments = pd.DataFrame(
-            np.full((n_non_fixed_experiments, len(domain.inputs)), None),
-            columns=column_keys,
-        )
-
-        one_set_of_experiments.mask(
-            binary_fixed_experiments.notnull(),
-            other=binary_fixed_experiments,
-            inplace=True,
-        )
-
-        if partially_fixed_experiments is not None:
-            one_set_of_experiments.mask(
-                partially_fixed_experiments.notnull(),
-                other=partially_fixed_experiments,
-                inplace=True,
-            )
-
-        if fixed_experiments is not None:
-            one_set_of_experiments = pd.concat(
-                [fixed_experiments, one_set_of_experiments]
-            ).reset_index(drop=True)
-
-        if sampling is not None:
-            sampling.loc[:, list_keys] = one_set_of_experiments[list_keys].to_numpy()
-
-        # minimizing with the current fixation
-        try:
-            current_design = find_local_max_ipopt(
-                domain,
-                model_formula,
-                n_experiments,
-                delta,
-                ipopt_options,
-                sampling,
-                None,
-                one_set_of_experiments,
-                objective,
-            )
-            domain.validate_candidates(
-                candidates=current_design.apply(lambda x: np.round(x, 8)),
-                only_inputs=True,
-                tol=1e-4,
-                raise_validation_error=True,
-            )
-            temp_value = objective_class.evaluate(
-                current_design.to_numpy().flatten(),
-            )
-            if minimum is None or minimum > temp_value:
-                minimum = temp_value
-                optimal_design = current_design
-            if verbose:
-                print(
-                    f"branch: {i} / {len(all_n_fixed_experiments)}, "
-                    f"time: {time.time() - start_time},"  # type: ignore
-                    f"solution: {temp_value}, minimum after run {minimum},"
-                    f"difference: {temp_value - minimum}"
-                )
-        except ConstraintNotFulfilledError:
-            if verbose:
-                print("skipping branch because of not fulfilling constraints")
-    return optimal_design
 
 
 def find_local_max_ipopt(
